Java version, Input, scroll and timing target
Java version was added to the pom in order to keep it updated, the Input class was also added, this class creates a custom good looking text input, the ScrollPane as well creates a better looking scrollpane, the PublicSetter class creates a timing target that doesn't rely on public getters and setters. Some comments were fixed in the Window class as well than the imports.